Refactor code to remove react-redux, prop-types, material-ui and redundant files. Move utils files around.
