--- src/pages/redux/modules/reducer.js
@@ -1,6 +1,0 @@
-import { combineReducers } from "redux";
-
-import wishlist from "./wishlist";
-import scheduled from "./courseSelect";
-
-export default combineReducers({ wishlist, scheduled });
